fixed the copyright year issue,renaming validEntityName to valdiateEntityName and made validateEntityName private to package services
