+++ menas/src/main/scala/za/co/absa/enceladus/rest/services/VersionedModelService.scala
@@ -86,7 +86,7 @@ abstract class VersionedModelService[C <: VersionedModel](versionedMongoReposito
     }
   }
 
-  def validateEntityName(name: String , entityType:String ): Unit  = {
+  private[services] def validateEntityName(name: String , entityType:String ): Unit  = {
     if (name.matches("""\w+""")) name
       else throw new Exception(s"$entityType name must not contain whitespace : $name")
   }
