+++ menas/ui/components/schema/schemaMain.controller.js
@@ -123,7 +123,7 @@ sap.ui.controller("components.schema.schemaMain", {
 					"Schema name '" + schema.name + "' already exists. Choose a different name.")
       isOk = false;
 		}
-      else if (GenericService.validEntityName(schema.name)) {
+      else if (GenericService.validateEntityName(schema.name)) {
       sap.ui.getCore().byId("newSchemaName").setValueState(sap.ui.core.ValueState.Error)
       sap.ui.getCore().byId("newSchemaName").setValueStateText(
           "Schema name '" + schema.name  + "' should not have spaces. Please remove spaces and retry")
