+++ menas/ui/service/GenericService.js
@@ -105,7 +105,7 @@ var GenericService = new function() {
         fnHelper("", sModelAcc, paths)
     };
 
-    this.validEntityName = function(sName) {
+    this.validateEntityName = function(sName) {
       if (/\W/.test(sName))
           return true;
       else return false;
